+++ onnx/ops.go
@@ -162,6 +162,22 @@ func getIntAttrOr(node *protos.NodeProto, attrName string, defaultValue int) int
 	return int(attr.I)
 }
 
+// getDTypeAttrOr gets a int attribute for node if present and convert to a GoMLX dtype, or return the given defaultValue.
+// It panics with an error message if the attribute is present but is of the wrong type.
+func getDTypeAttrOr(node *protos.NodeProto, attrName string, defaultValue dtypes.DType) dtypes.DType {
+	attr := getNodeAttr(node, attrName, false)
+	if attr == nil {
+		return defaultValue
+	}
+	assertNodeAttrType(node, attr, protos.AttributeProto_INT)
+	onnxDType := protos.TensorProto_DataType(int32(attr.I))
+	dtype, err := dtypeForONNX(onnxDType)
+	if err != nil {
+		exceptions.Panicf("unsupported ONNX data type %q for attribute %q in %s", onnxDType, attrName, nodeToString(node))
+	}
+	return dtype
+}
+
 // getBoolAttrOr gets a boolean attribute (ONNX uses an int value of 0 or 1) for node if present or return the given defaultValue.
 // It panics with an error message if the attribute is present but is of the wrong type.
 func getBoolAttrOr(node *protos.NodeProto, attrName string, defaultValue bool) bool {
@@ -1022,7 +1038,7 @@ func convertMax(operands []*Node) *Node {
 //
 ////////////////////////////////////////////////////////////////////
 
-// convertLSTM converts a ONNX node to a GoMLX node.
+// convertLSTM converts an ONNX node to a GoMLX node.
 //
 // The GoMLX version used ONNX version as inspiration, so they have the same feature support.
 //
@@ -1124,4 +1140,56 @@ func convertLSTM(m *Model, convertedOutputs map[string]*Node, node *protos.NodeP
 	}
 
 	return allHiddenStates
+}
+
+////////////////////////////////////////////////////////////////////
+//
+// Quantization related ops.
+//
+////////////////////////////////////////////////////////////////////
+
+// convertDequantizeLinear converts the corresponding ONNX node to a GoMLX node.
+//
+// Not yet supporting axis or block dequantization.
+//
+// See ONNX documentation in:
+// https://onnx.ai/onnx/operators/onnx__DequantizeLinear.html
+func convertDequantizeLinear(nodeProto *protos.NodeProto, inputs []*Node) *Node {
+	// Attributes:
+	// - Axis (optional) on which to apply the multi-valued scale.
+	// - blockSize: optional, only active if != 0. Not yet implemented.
+	targetAxis := getIntAttrOr(nodeProto, "axis", 1)
+	blockSize := getIntAttrOr(nodeProto, "blockSize", 0)
+	if blockSize != 0 {
+		exceptions.Panicf("DequantizeLinear: support for attribute 'block_size' is not yet implemented")
+	}
+	outputDType := getDTypeAttrOr(nodeProto, "output_dtype", dtypes.Float32)
+
+	x := inputs[0]
+	scale := inputs[1]
+	if !scale.IsScalar() {
+		// Add extra axes of dim=1 in scale to match x's rank.
+		if scale.Rank() != 1 {
+			exceptions.Panicf("DequantizeLinear: scale must be a scalar or 1D, got %s instead", scale.Shape())
+		}
+		newScaleShape := x.Shape().Clone()
+		for axis := range newScaleShape.Dimensions {
+			if axis != targetAxis {
+				newScaleShape.Dimensions[axis] = 1
+			} else if newScaleShape.Dimensions[axis] != scale.Shape().Dimensions[0] {
+				exceptions.Panicf("DequantizeLinear: scale must have same dimension as the input axis %d (input shape=%s), got %s instead", targetAxis, x.Shape(), scale.Shape())
+			}
+		}
+		scale = Reshape(scale, newScaleShape.Dimensions...)
+	}
+
+	if len(inputs) == 3 {
+		xZeroPoint := inputs[2]
+		x = Sub(x, xZeroPoint)
+	}
+	x = Mul(ConvertDType(x, scale.DType()), scale)
+	if x.DType() != outputDType {
+		x = ConvertDType(x, outputDType)
+	}
+	return x
 }