+++ docs/CHANGELOG.md
@@ -1,3 +1,9 @@
+# Next 
+
+* Added support to other ONNX dtypes that requires conversion during reading.
+  Also added conversion when saving values back to the ONNX proto.
+* Updated dependencies to GoMLX.
+
 # v0.2.3 2025/05/31
 
 * Added Save/Check values of outputs for internal/benchmarks: allows it to be